test/requirements/pypi.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras pypi.in
#
--only-binary :all:

ansi2html==1.9.2
    # via -r pypi.in
execnet==2.1.1
    # via pytest-xdist
iniconfig==2.0.0
    # via pytest
packaging==24.1
    # via pytest
pluggy==1.5.0
    # via pytest
psutil==6.0.0
    # via pytest-xdist
py==1.11.0
    # via pytest-html
pytest==8.3.3
    # via
    #   -r pypi.in
    #   pytest-html
    #   pytest-metadata
    #   pytest-xdist
pytest-html==3.2.0
    # via -r pypi.in
pytest-metadata==3.1.1
    # via pytest-html
pytest-xdist==3.6.1
    # via -r pypi.in
pyyaml==6.0.2
    # via -r pypi.in
setproctitle==1.3.3
    # via pytest-xdist
